requirements.txt
# Core AI and API dependencies
google-generativeai>=0.3.0

# Web interface
streamlit>=1.28.0

# Image processing
Pillow>=9.0.0
numpy>=1.21.0

# Additional utilities
requests>=2.28.0

# Development and testing (optional)
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0 